+++ CHANGELOG.md
@@ -8,14 +8,17 @@ GAMSPy 1.0.2
   - Support output in Container constructor.
   - Fix debugging_level bug.
   - Fix singleton assignment bug.
+  - Allow an alternative syntax for variable/equation attributes (e.g. b[t].stage = 30).
 - Testing
   - Remove license uninstall test to avoid crashing parallel tests on the same machine.
   - Add tests for the generated solve strings for different type of problems.
   - Add a test for Container output argument.
   - Add tests for debugging_level.
+  - Add tests for the alternative syntax for variable/equation attributes.
 - Documentation
   - Fix broken links in the documentation.
   - Add a ci step to check doc links.
+  - Improve the wording of debugging document.
 
 GAMSPy 1.0.1
 ------------